+++ auto_database_backup/doc/RELEASE_NOTES.md
@@ -29,3 +29,11 @@
 #### UPDT
 
 - Fixed internal server error in Onedrive and Google Drive integration.
+
+## Module <auto_database_backup>
+
+#### 18.04.2024
+#### Version 16.0.5.0.1
+#### UPDT
+
+- Fixed the error while inputting list_db = False in odoo conf file.